requirements.txt
opencv-python
mediapipe
numpy
pandas
scikit-learn
joblib
pillow
tk
